Update BWAPI, BWTA, and BWSAL
Everything is built from subversion, you can also download
the files from the download section of git hub, which also
includes the source files.

Since this includes a new BWAPI.dll, this needs to be copied to
your chaoslauncher folder, or I also added the dll's to the revision
folder, so you could just copy those over and leave your chaoslauncher
folder the same. You might want to do this anyways, since I added some
more dll's to the revision folder.

You might have to build clean on your solution.

This was tested for both the release and debug BasicAIModule.