Add BWTA analyze and screen graphics
Copied EnhancedUI files from newer BWSAL BasicAIModule,
except commented out some functions as they may not be
set up for our agents.  So mainly just uses BWTA.

Add Util Foreach.h header file as BWSAL uses this if we
decide to add any code from it, so if you see 'foreach'
instead of 'for each', then you know that header is being
used.  Looks like it's just a wrapper to support GNUC.

Note it takes a second or 2 to analyze the map in onStart(),
so that is why it hangs for a bit.