add EnhancedSide for EnhancedChokepoint
EnhancedChokepoint now stores different sides of
it's bounding box based for each of it's 2 regions.
The sides are added if each side lies in the region.
Basically, this tells us if our base lies in a certain
BWTA region, which sides of the chokepoint bounding box
would I like to wall off if need be or which sides am I
most concerned with.  Ideally you want to wall off the
inside of the chokepoint not the outside, now we know
which is the inside.

The red line represents the inside, the purple boxes
are the tiles that contain the end points of the chokepoint
line, the yellow box is the tile that contains the center
of the line.

I also change the base region outline to have a different
color.

Finally, I found that the BWTA isInside() for polygon
does not work all the time :(  So I found some better
code online and added that with a reference.