Building Barracks
Agent.h:  Added a buildingReserved bool to tell whether space has been reserved.
This is because the BWAPI function isConstructing() does not return true immediately,
leading to repeated calls to BFSBuildingPlacer.

  Also, added buildingLocation for the return from BFSBuildingPlacer.

SCVAgent:  Change the build logic to reflect the new buildingReserved bool.

CombatManager:  Builds a barracks with SCVs when it has them.

TODOS:
* Need to enforce 1-1 mapping from Agents to Managers
* Need a means of reserving resources (minerals)
* Managers that build structures need to know how many they have built, plan to build,
  etc. to prevent overconstruction.