Update strategizer, supply M, scvAgent, add gas M
Changed startegizer to only assign new agents
to default managers, then agents are reassigned
based upon need

Changed supply manager to only have one scv
and to only build when needed

Updated build state for SCVAgent, so it saves
the build location to itself and then when the
build command goes through (takes a while because
resources are not enough) it builds at the known
location

Also added idle state for scv agent, so that it
won't keep building new buildings in the build
state (we don't always want that).  This allows
for control back to the manager

Set up code for the gas manager, not working yet.

Add special building place for TacticalBuildingPlacer
for the gas refinery.