Add combatManager back in
Again the combatManager gets 1 scv for now
and it builds 2 barracks.

Overall the bot seems to run a little smoother
now.

Managers definitely need to be built out more
though.

The build queue with bids would be very helpful
as well.

TacticalBuildingPlacer should be improved so that
in can track building locations per region, so it
can put extra supply depots in a row and it still
needs to be set up for walling and placing bunkers