Chokepoint search working
Fixed all the bugs I could find, it doesn't crash anymore by
getting asserts on the vector out of range.  Search seems to work
well, on benzene map for one of the start locations, the tiles are
off into the other region.  I am guessing this is because if no
buildable tiles are found in a round, the search expands out normally
as if all the tiles from the last round were found to be buildable.
This could also be changed to don't expand the number of tiles,
just move outward.

The issue of finding buildable tiles in the wrong region, this can
be added in as another check.

Currently all the buildable tiles that were found while searching
for a build tile are in white, the actual build tile is one of
those, haven't changed the color yet.

So in effect the search only finds the first tile that can be
used for constructing.  You would then have to search off of that
location to find the optimal pattern for placing the depots or
bunkers.