Draw supply depot locations for chokepoint
So for it seems like these are pretty close to the chokepoint.

Note: it is possible for 2 chokepoints that are close together
to come up with overlapping regions. Hopefully we will be able
to now do a larger search with just canBuildHere and not worry
about buildable tiles by themselves, and then once the build
pattern is established, reserve the tilse. The TacticalBuildPlacer
can do this higher level planning and then that would fix the
overlapping as well.