Add chokepoint search in TacticalBuildingPlacer
Add chokepointBuildPatternSearch() to TBP
Add checkCoverage() and checkGap() to EnhancedSide

Basically TBP is now searching for optimal patterns
to wall off chokepoints with Supply Depots.  This is
inherantly trickys since chokepoints come in all shapes
and sizes.  So as of now it can find the most basic patterns
and then more functionality will be added for it to search
for special cases.  It will at least give one location for a
building close to the chokepoint if it can not find an optimal
pattern.

The checkCoverage() and checkGap() functions help to assess
how effective the build patterns are.  checkCoverage() helps
to see if the chokepoint is actually being blocked and checkGap()
finds the minimum gap at the side ends of the supply depot wall.