Update chokepoint walling
Main changes:

Using side end, facing side gaps, and chokepoint coverage to
determine build patterns

Gaps are measured in walkable tiles which are 8x8 pixel tiles.
This allows for a more fine grained test instead of using
buildable tiles which are 32x32 px.  Also isWalkable() checks
static map data, so this will not be influenced by units on the map.

There are 3 ways to grow out the wall of supply depots, they are
grow by coverage, grow by facing side gap, and grow by length. So
all are associated with a check to make sure the growth is
making an improvement.

Shifting and reverse grow shifting are still limited.  Nothing is
in place for staggered building (Astral Balance Map).

Tested on the 3 2-player AIIDE maps.