Build a bunker near chokepoint
Uses the build tile for the first enhanced chokepoint found
for the base starting location region.  These build tiles
are setup for supply depots, however there is no point in
building a wall if you have nothing to back it up with. This
will allow us to get the bunker and squad defense logic written.

May want to add a Bunker agent, however not sure where this will go.
Structure agent is currently setup to train units, and a bunker
can't produce any agents.

May be good to move training logic to a producer agent and keep
structure agent for basic building logic.