Marines in bunker via bunker squad under CM
Combat manager now has 3 sets of squads, attack, defense, and bunker.
Currently the defend squad is a temp holder until the bunker is built
and then the marines are moved from that squad to a bunker squad.
Upon entering the bunker squad, if a bunker target is assigned, then
they automatically load themselves in it.

Bunker agent is very minimal right now, but I'm sure more functionality
will be needed for tracking repairs and other things of that nature.
Bunker agent is also under the combat manager, not the build manager.

Right now, one less marine is being loaded into the bunker, it could
still be stuck in the defend squad, not exactly sure yet...