Attack squads only attack when at full force
Use AttackSquadSize to set when to attack with squad
and when to make new squads.  Its setup to be variable
and its setup to increase each time another squad is setup.

Now squads move faster since they are not waiting for new
agents once they have already left to attack the enemy
base.