Add enemy unit targets in range to attack squads
Combat manager sets the unit targets for the leader
based on weakest enemy, this should be changed to
closest weakest enemy.

Actor agent attack state is set up to attack the
unit target if it's in range, otherwise start moving
towards the unit targets position, or if there isn't
a unit target, then go to the set position target.

So if no unit targets are found, then the unit target
is reset to NULL allowing for target positions to take
precedence.

Enemy unit targets are highlited by a green / yellow circle
just like the squad leader is highlighted.

Also invalid target position checking was added to the
attack state, as it was found that some targets would
end up off the map, leading everyone into a corner.