Add buildQueue to agents.  Used to queue build orders.
BuildManager now has an additional parameter for build: count.  This can be used to specify
how many of a UnitType to build.

Agent.build() now adds the specified count to the Agent's build queue.

Agents train/build only as many of the UnitType as specified in buildQueue, rather than
previous behavior which was to build indefinitely.